+++ routes/blogs.js
@@ -1,8 +1,9 @@
-var express = require('express');
-var router = express.Router();
-var bodyParser = require('body-parser');
-var path = require('path');
-var moment = require('moment');
+const express = require('express'),
+      router = express.Router(),
+      bodyParser = require('body-parser'),
+      moment = require('moment'),
+      middleware = require('../middleware')
+      User = require('../model/user');
 
 //ย้ายรูปจาก form หน้า editprofile ไปเก็บในโฟลเดอร์ images/img-profile
 var multer = require('multer');
@@ -31,14 +32,7 @@ var StorageOfimagepost = multer.diskStorage({
 var upload_imgpost = multer({storage : StorageOfimagepost});
 
 
-//passport login
-var passport = require('passport');
-//import model เข้ามาใช้แค่ในระบบ login
-var User = require('../model/usermodel');
-//ตรวจข้อมูลและการแจ้งข้อผิดพลาด ใช้ในการหน้า register
-var { check, validationResult } = require('express-validator');
-//login 
-var LocalStrategy = require('passport-local').Strategy;
+
 //connect DB
 var mongoose = require('mongoose');
 var ObjectId = require('mongodb').ObjectId;
@@ -74,7 +68,7 @@ let CatelogSchema = new mongoose.Schema({
 let conCatelog = mongoose.model("categories", CatelogSchema);
 
 //แสดงหน้าแรก ถ้า login แล้วจะแสดงอีกหน้านึ่ง
-router.get('/',checkAuthentication, async function(req, res,) {
+router.get('/', async function(req, res,) {
   //ไปดึงข้อมูล posts มาแสดงหน้าแรก
     const songkran_post = await conPost.find({category : "เที่ยวสงกรานต์"}).limit(4);
 
@@ -85,145 +79,17 @@ router.get('/',checkAuthentication, async function(req, res,) {
     const cat = await conCatelog.find();
 
     res.render("index",{ section1 : songkran_post, Marketfloat : marketfloat_post, Category : cat});
-  });
-
-
-  function checkAuthentication(req,res,next){
-    //ตรวจสอบว่า login แล้วหรือยัง
-    if(req.isAuthenticated())
-    {
-      //ถ้า login แล้ว ให้ next() คือ res.render("index") ออกไปเลย
-      return next();
-    } 
-    else
-    {
-      //กรณีถ้ายังไม่ login จะ redirect ไปหน้าไหนก่อน
-      //next() -> เข้าหน้าแรกได้เลย ไม่loginก็ได้
-      return next();
-    }
-  }
-
-router.get("/register", function(_req, res)
-{
-  res.render("register");
 });
 
-router.post("/register",[
-  check("username","กรุณาป้อนชื่อผู้ใช้").not().isEmpty(),
-  check("email","กรุณาป้อนอีเมล").isEmail(),
-  check("password","กรุณาป้อนรหัสผ่าน").not().isEmpty(),
-  check("password2","กรุณายืนยันรหัสผ่าน").not().isEmpty()
-  ] ,function(req, res)
-  {
-    const result = validationResult(req);
-    var errors = result.errors;
-    if(!result.isEmpty())
-    {
-      //return to page
-      res.render("register", { errors : errors });
-    }
-    else
-    {
-      let username = req.body.username;
-      let email = req.body.email;
-      let password = req.body.password;
-      let password2 = req.body.password2;
-
-      //ค่าเริ่มต้นในการ register แล้วค่อยไปเพิ่มข้อมูลอื่นๆ ในภายหลัง
-      var newUser = new User
-      ({
-        username:username,
-        email:email,
-        password:password,
-        image: "no-imgprofile.png"
-      })
-      User.createUser(newUser,function(err){
-        if(err) console.log(err);
-        else
-        {
-          res.location('/blogs/');
-          res.redirect('/blogs/login');
-        }
-      });
-      
-    }
-  });
 
 router.get("/blogs1", function(req, res){
   res.render("index1");
 });
 
-router.get("/login", function(_req, res)
-{
-  res.render("login");
-});
 
-router.get("/logout", function(req, res)
-{
-  //ทำลาย session ทิ้ง
-  req.logout();
-  res.redirect("/blogs");
-});
 
-router.post("/login", passport.authenticate('local',{
-  // ถ้าไม่สำเสร็จให้ไปที่หน้า /blogs/login
-  failureRedirect:'/blogs/login',
-  failureFlash:false
-}),
-function(req, res)
-{
-  // สำเร็จ
-  res.redirect('/blogs');
-});
 
-passport.serializeUser(function(user,done)
-{
-  done(null,user.id);
-});
-
-passport.deserializeUser(function(id,done)
-{
-  //User อันนี้ เป็น module ที่ export ออกมาจาก /model/usermodel.js แล้วเข้าไปใช้งาน method ชื่อว่า getUserById
-  User.getUserById(id,function(err,username)
-  {
-    done(err,username)
-  })
-});
-
-passport.use(new LocalStrategy(function(email,password,done)
-{
-  //ค้นหาด้วย email 
-  User.getUserByEmail(email,function(err,user){
-    //ส่งข้อมูล user กลับมา
-    if(err) throw errors
-    if(!user)
-    {
-      //อีเมลไม่ถูกต้อง ไม่พบผู้ใช้
-      return done(null,false)
-    }
-    else
-    { 
-      //อีเมลล์ถูกค้อง แล้วค่อยเปรียบเทียบ password
-      User.comparePassword(password,user.password,function(err,isMatch)
-      {
-        if(isMatch)
-        {
-          //รหัสผ่านถูกต้อง
-          //return ออกไปเก็บ session ตัวแปร locals.user สามารถเรียกใช้งานได้ทั้งระบบ
-          return done(null,user);
-        }
-        else
-        {
-          return done(null,false)
-        }
-        
-      })
-      //ส่งไปหา passport.serializeUser
-    }
-  });
-}));
-
-router.get("/new",checkAuthentication,async function(req, res)
+router.get("/new",middleware.checkAuthentication,async function(req, res)
 {
   const cat = await conCatelog.find();
   res.render("Addpost",{ categories : cat });
@@ -273,7 +139,7 @@ router.get("/review/:id", async function(req, res)
       res.render("review",{ Blogs : postreview , Category : cat, moment : moment});
 });
 
-router.get("/edit/:postid",async function(req, res){
+router.get("/edit/:postid",middleware.checkAuthentication,async function(req, res){
   const { postid } = req.params;
   const Editpost = await conPost.findById(postid);
   const cat = await conCatelog.find();
@@ -313,7 +179,7 @@ router.get("/delete/:postid",async function(req, res){
   await conPost.remove({ _id:postid });
 });
 
-router.get("/profile/id=:id", async function(req, res){
+router.get("/profile/id=:id",middleware.checkAuthentication, async function(req, res){
   const { id } = req.params;
   const result = await conUser.aggregate(
   [
@@ -368,7 +234,7 @@ router.post("/profile/edit/id=:userid", upload_profile.single('imgprofile'), asy
   }
 });
 
-router.get("/profile/edit/id=:id", async function(req, res)
+router.get("/profile/edit/id=:id",middleware.checkAuthentication, async function(req, res)
 {
   res.render("Editprofile");
 });
