+++ views/layout/layoutnavbar.ejs
@@ -26,7 +26,7 @@
                   <a class="dropdown-item" href="/blogs/profile/edit/id=<%= user._id %>"><i class="fas fa-user-edit"></i> แก้ไขโปรโฟล์</a>
                   <a class="dropdown-item" href="/blogs/profile/id=<%= user._id %>"><i class="far fa-list-alt"></i> แก้ไขบทความ</a>
                   <a class="dropdown-item" href=""><i class="far fa-bookmark"></i> ที่บันทึกไว้</a>
-                  <a class="dropdown-item" href="/blogs/logout"><i class="fas fa-sign-out-alt"></i> ออกจากระบบ</a>
+                  <a class="dropdown-item" href="/logout"><i class="fas fa-sign-out-alt"></i> ออกจากระบบ</a>
                 </div> 
               </div>
         <% } %> 
@@ -35,7 +35,7 @@
                 <li class="nav-item active"><a class="nav-link" href="">หน้าแรก</a></li>
             </ul>
             <span>
-                <a href="/blogs/login">
+                <a href="/login">
                     <button type="button" class="btn btn-primary"><i class="fas fa-user"></i> ลงชื่อเข้าใช้</button>
                 </a>
             </span>
