+++ views/login.ejs
@@ -47,7 +47,7 @@
 <body class="background-login">
 	<div >	
 		<div class="container">
-			<form class="form-group form-login" method="POST" action="/blogs/login"> 
+			<form class="form-group form-login" method="POST" action="/login"> 
 				<h1 class="text-center mt-lg-0 mt-md-4 mt-sm-4 ">Login</h1>
 				<div class="form-group">
       				<label for="Email">Email:</label>
